Update the marshaller to be faster at sorting indexed objects.
Prior to this, each comparison would try to get the index by converting the object to the IndexGettable interface which is a relatively expensive process. This does the conversion earlier in the process. The downside is that it involves creating a temporary slice of the intermediate indexedObject type and then copying it again.